local_test/requirements.txt
beautifulsoup4==4.13.4
jieba==0.42.1
Jinja2==3.1.6
matplotlib==3.10.1
modelscope==1.25.0
PyMySQL==1.1.0
numpy==1.26.4
requests==2.31.0
sanic==23.12.1
wordcloud==1.9.4
addict==2.4.0
datasets==2.16.0
fake_useragent==2.2.0
httpx==0.26.0
PyJWT==2.8.0
Pillow==10.2.0
werkzeug==3.0.1
python-jose==3.3.0